+++ src/main/java/wooteco/subway/service/dto/request/LineServiceRequest.java
@@ -6,10 +6,10 @@ public class LineServiceRequest {
     private String color;
     private Long upStationId;
     private Long downStationId;
-    private int distance;
+    private Long distance;
     private Long extraFare;
 
-    public LineServiceRequest(Long id, String name, String color, Long upStationId, Long downStationId, int distance,
+    public LineServiceRequest(Long id, String name, String color, Long upStationId, Long downStationId, Long distance,
                               Long extraFare) {
         this.id = id;
         this.name = name;
@@ -20,11 +20,11 @@ public class LineServiceRequest {
         this.extraFare = extraFare;
     }
 
-    public LineServiceRequest(Long id, String name, String color, Long upStationId, Long downStationId, int distance) {
+    public LineServiceRequest(Long id, String name, String color, Long upStationId, Long downStationId, Long distance) {
         this(id, name, color, upStationId, downStationId, distance, null);
     }
 
-    public LineServiceRequest(String name, String color, Long upStationId, Long downStationId, int distance) {
+    public LineServiceRequest(String name, String color, Long upStationId, Long downStationId, Long distance) {
         this(null, name, color, upStationId, downStationId, distance);
     }
 
@@ -52,7 +52,7 @@ public class LineServiceRequest {
         return downStationId;
     }
 
-    public int getDistance() {
+    public Long getDistance() {
         return distance;
     }
